backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
SQLAlchemy==2.0.30
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
requests==2.32.3
email-validator==2.1.1
